requirements.txt
gitpython>=3.1.40
radon>=6.0.1
pylint>=3.0.0
pydriller>=2.6
tabulate>=0.9.0
jinja2>=3.1.2
pyyaml>=6.0.1
click>=8.1.7
